Implemented the L1 Basic Iterative Attack
I had problems implementing the Fast Gradient Sign Attack from Foolbox, since one function in the middle of it wasn't implemented, and the attack doesn't work because of that. Instead, I implemented the L1 Basic Iterative Attack; to make it fully operational with the data we have right now, a function in one file needs to be changed from "cross_entropy" to "binary_cross_entropy", since the creators didn't expect conducting the attack on models with binary results. The function in question returns an error when running without the change, so it shouldn't be difficult to find and change. Models with non-binary results should work with the code as it is.