Salt and Pepper + refactoring
Added an attack - "Salt and Pepper noise attack" - and changed the way arguments are structured in attacks with two parents - for now, that means L1BasicIterative and SaltAndPepperNoise. If an attack has two parents, one is a class from a specific library (in this case, foolbox) and the other is our generic wrapper for that library, then the __init__ function of that attack should accept two sets of parameters - one for the __init__ function of each parent. Previously, the specific parent only used default parameters.